fix: arrow missing from select box
