+++ libs/ui-toolkit/src/components/input/input.tsx
@@ -3,10 +3,7 @@ import { forwardRef } from 'react';
 import classNames from 'classnames';
 import type { IconName } from '../icon';
 import { Icon } from '../icon';
-import {
-  includesLeftPadding,
-  includesRightPadding,
-} from '../../utils/class-names';
+import { inputClassNames } from '../../utils/form-elements';
 
 type InputRootProps = InputHTMLAttributes<HTMLInputElement> & {
   hasError?: boolean;
@@ -59,37 +56,6 @@ type InputAppend = NoPrepend &
 type AffixProps = InputPrepend | InputAppend;
 
 type InputProps = InputRootProps & AffixProps;
-
-export const inputClassNames = ({
-  hasError,
-  className,
-}: {
-  hasError?: boolean;
-  className?: string;
-}) => {
-  return classNames(
-    [
-      'appearance-none',
-      'flex items-center w-full',
-      'box-border',
-      'border rounded-none',
-      'bg-clip-padding',
-      'border-black-60 dark:border-white-60',
-      'bg-black-25 dark:bg-white-25',
-      'text-black placeholder:text-black-60 dark:text-white dark:placeholder:text-white-60',
-      'text-ui',
-      'focus-visible:shadow-focus dark:focus-visible:shadow-focus-dark',
-      'focus-visible:outline-0',
-      'disabled:bg-black-10 disabled:dark:bg-white-10',
-    ],
-    {
-      'pl-8': !includesLeftPadding(className),
-      'pr-8': !includesRightPadding(className),
-      'border-vega-pink dark:border-vega-pink': hasError,
-    },
-    className
-  );
-};
 
 export const inputStyle = ({
   style,
@@ -165,7 +131,7 @@ export const Input = forwardRef<HTMLInputElement, InputProps>(
     const hasPrepended = !!(prependIconName || prependElement);
     const hasAppended = !!(appendIconName || appendElement);
 
-    const inputClassName = classNames('h-28', className, {
+    const inputClassName = classNames('appearance-none', 'h-28', className, {
       'pl-28': hasPrepended ?? hasAppended,
     });
